fmt_human: Improve types
size_t may not be enough, use 64 bit integers

+++ util.c
@@ -1,6 +1,7 @@
 /* See LICENSE file for copyright and license details. */
 #include <errno.h>
 #include <stdarg.h>
+#include <stdint.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
@@ -93,7 +94,7 @@ bprintf(const char *fmt, ...)
 }
 
 const char *
-fmt_human(size_t num, int base)
+fmt_human(uint64_t num, int base)
 {
 	double scaled;
 	size_t i, prefixlen;
