+++ util.h
@@ -1,4 +1,6 @@
 /* See LICENSE file for copyright and license details. */
+#include <stdint.h>
+
 extern char buf[1024];
 
 #define LEN(x) (sizeof (x) / sizeof *(x))
@@ -10,5 +12,5 @@ void die(const char *, ...);
 
 int esnprintf(char *str, size_t size, const char *fmt, ...);
 const char *bprintf(const char *fmt, ...);
-const char *fmt_human(size_t num, int base);
+const char *fmt_human(uint64_t num, int base);
 int pscanf(const char *path, const char *fmt, ...);
